backend/requirements.txt
psycopg2-binary
fastapi
uvicorn
python-dotenv
openai
pydantic
